Georeferencing: Fix state switching
Normally, convergence and grid scale factor are calculated or reset
in updateGridTransformation. However, when switching from normal mode
(i.e. georeferenced) to local mode (no CRS), these variables must not
be reset in order to keep the transformation between map coordinates
and "projected" coordinates unchanged.
Consequently, we must save and reload the original grid scale factor,
and and initialize convergence, when dealing with local CRS.
In a sense, this implements what is displayed in the dialog:
Grivation and combined scale factor are "locked", either until
declination or auxiliary scale factor are edited manually, or
until a new CRS is set.
Note that there is no way to reset grid scale factor and convergence
except for choosing a new CRS. But there is little reason to keep
a map without CRS and still being interested in these parameters
at all.

The changed behaviour can be tested manually with world-file.xmap:
1. Change CRS to "Local".
2. Save and close.
3. Open again.
4. Change CRS back to EPSG 2180.
... and monitor:
- the EPSG 2180 coordinates of a corner point,
- the template positioning.