GeoreferencingDialog: Preserve centering of map and printer views
When the Georeferencing affects map coordinates of objects,
it is convenient if the map view and other views remain
centered on the same objects, rather than stay with the
map coordinates and move to other objects (or away from the
visible portion of the map entirely).

This commit adds support for this feature, to be enabled
in the "Realign map" dialog.