GeoreferencingDialog: Fewer grivation updates
This revises commit 3139987 from GH-1222 to update grivation only
when there's a valid CRS spec, because without a CRS spec the
convergence, which would be used to calculate grivation, has no basis.
For the same reason, the test is revised to not update grivation when
the state of the Georeferencing is Local.