GeoreferencingDialogs: Add ReferenceSystemDialog
Recast reference system...
ReferenceSystemDialog is for changes to the CRS and/or reference point with
minimal effect on georeferencing between map coordinates and geographic
coordinates. If the CRS is changed, then there's a report to the user of
change to the geographic coordinates of map objects. Changing the
reference point has no effect on geographic coordinates of objects, but
it does affect how declination is measured.

In the ReferenceSystemDialog, various items are hidden if the georeferencing
is in Local state. Visibility is managed through widgets, which is in
keeping with the coding style of other dialogs, though here it is still
quirky, especially in the introduction of the geographic_ref_widget for
the sole purpose of hiding everything in the geographic_ref_layout.

Order of reference point edit rows is in agreement with the reference point
edit rows in the GeoreferencingDialog. Parameters of the geographic aspect
are topmost in each pair of corresponding parameters, consistently for all
georeferencing dialogs and for all three parameter pairs.

This dialog's two modes are based on whether the georeferencing has a CRS.
The CRS selector is always present.

In the MapEditor the "Recast reference system..." menu items is disabled
when appropriate.