Map dialogs: Provide a full set of transformations
The menu commands are "Rotate all objects", "Scale all objects", and
"Shift all objects".
The menu name "Rotate map" is changed to "Rotate all objects".
The StretchMapDialog is updated to handle the "Scale all objects" menu.
The ShiftMapDialog is added to support "Shift all objects".

The option to not adjust the georeferencing reference point
is removed from the rotate and scale objects dialogs.

This commit downgrades a feature of the GeorefrencingDialog, in which
it would offer to stretch or shrink the map in proportion to a scale
factor change. The feature is to be removed in a subsequent commit.

Fixes to e531327 full set of transformations