Georeferencing: Add 'keep_georeferencing' option when setting reference point
Adds a 'keep_georeferencing' parameter to setMapRefPoint, setProjectedRefPoint,
and setGeographicRefPoint. These are for use by the "Recast reference system"
dialog. When the parameter is set, these changes to the reference point
have no effect on the map coordinates of objects.

Also, a boolean return is added to these methods as they may fail
due to the CRS.

The setGeographicRefPoint will not set the state to Normal.
To better support Local state, don't assume that just because the
latitude and longitude are present, that the user will need to set a CRS.