MapAlignmentDialog: Add scale edit and options
Combine features of the "Change map scale" dialog into this
"Realign map" dialog, primarily to provide for setting the scale,
as changing the scale normally is a realignment in that it affects
the appearance of objects on the map while leaving their
geolocations unchanged.

Also added are two options. The option for whether or not to align
non-georeferenced templates has belonged in this dialog from the
start. The option to scale symbols is relevant if the nominal scale
is being changed.