GeoreferencingDialogs: Reflect need for geo ref point / declination
The three georeferencing dialogs (Set georeferencing, Recast reference system,
and Realign map) are enhanced to expose to the user whether the geographic
reference point has been actively set, and likewise for the declination.
Showing "no value" in their edit boxes makes clear what the user must supply
prior to a successful click on "OK".  This affects the GeoreferencingDialog
and ReferenceSystemDialog.

In the MapAlignmentDialog, declination editing is disabled
if there is no geographic reference point.

The menu item for ReferenceSystemDialog is disabled if there
is no geographic reference point. Also, there must either be
a valid CRS or a declination value.

When the CRS is -local- in the ReferenceSystemDialog, the CRS selector
is normally enabled but reference point widgets are disabled. However,
regardless of the the CRS selection, the selector cannot be modified if
'has_declination' is false.