Georeferencing: Add 'update_parameters' option when setting CRS
If UpdateGridParameter is passed to setProjectedCRS, the projected
reference point is updated to be consistent with the existing geographic
reference point, and grivation is updated similarly. As a result,
the effect on the geographic coordinates of map objects is minimal.

If UpdateGeographicParameter is passed to setProjectedCRS, the geographic
reference point is updated to be consistent with the existing projected
reference point, and declination is updated similarly.
Calling setProjectedCRS with UpdateGeographicParameter differs from calling it
with NoUpdate and subsequently calling setProjectedRefPoint. When
setProjectedRefPoint updates the geographic reference point, that is assumed to
be the user's intent. However, when the user's action is to set the CRS,
the existing projected reference point is most likely invalid as a basis
for updating the geograhic reference point.

The default value for 'update_parameters' is NoUpdate, which usually leaves
the georeferencing parameters inconsistent.